# c1one/c1one-front/docs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Uniqram Swagger",
    "description": "인스타그램 클론 프로젝트 API 명세서",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "JWT": []
    }
  ],
  "paths": {
...
    "/api/posts": {
      "post": {
        "tags": ["post-controller"],
        "operationId": "createPost",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponsePostResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/posts/{postId}/like": {
      "post": {
        "tags": ["post-controller"],
        "operationId": "likePost",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PostLikeResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/signin": {
      "post": {
        "tags": ["auth-controller"],
        "operationId": "signin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SigninRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "tags": ["auth-controller"],
        "operationId": "refreshToken",
        "parameters": [
          {
            "name": "refresh_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": ["auth-controller"],
        "operationId": "logout",
        "parameters": [
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Cookie"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/join": {
      "post": {
        "tags": ["auth-controller"],
        "operationId": "signup",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/posts/{postId}": {
      "get": {
        "tags": ["post-controller"],
        "operationId": "getPostDetail",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PostDetailResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["post-controller"],
        "operationId": "deletePost",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": ["post-controller"],
        "operationId": "updatePost",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/posts/profile/{userId}": {
      "get": {
        "tags": ["post-controller"],
        "operationId": "getUserPosts",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageUserPostResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "PostCreateRequest": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "mediaUrls": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "PostResponse": {
        "type": "object",
        "properties": {
          "postId": {
            "type": "integer",
            "format": "int64"
          },
          "content": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "mediaUrls": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "memberId": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          }
        }
      },
      "SuccessResponsePostResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/PostResponse"
          }
        }
      },
      "PostLikeResponse": {
        "type": "object",
        "properties": {
          "postId": {
            "type": "integer",
            "format": "int64"
          },
          "liked": {
            "type": "boolean"
          },
          "likeCount": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "SigninRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "Cookie": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "value": {
            "type": "string"
          },
          "attributes": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "path": {
            "type": "string"
          },
          "comment": {
            "type": "string",
            "deprecated": true
          },
          "version": {
            "type": "integer",
            "format": "int32",
            "deprecated": true
          },
          "maxAge": {
            "type": "integer",
            "format": "int32"
          },
          "secure": {
            "type": "boolean"
          },
          "httpOnly": {
            "type": "boolean"
          },
          "domain": {
            "type": "string"
          }
        }
      },
      "SignupRequest": {
        "required": ["confirmPassword", "password", "username"],
        "type": "object",
        "properties": {
          "username": {
            "maxLength": 20,
            "minLength": 2,
            "type": "string"
          },
          "password": {
            "maxLength": 100,
            "minLength": 6,
            "type": "string"
          },
          "confirmPassword": {
            "type": "string"
          }
        }
      },
...
      "PostUpdateRequest": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "remainImageUrls": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "newImageUrls": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "CommentListResponse": {
        "type": "object",
        "properties": {
          "commentId": {
            "type": "integer",
            "format": "int64"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "userName": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "likeCount": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "parentCommentId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "LikeUserDto": {
        "type": "object",
        "properties": {
          "postId": {
            "type": "integer",
            "format": "int64"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          }
        }
      },
      "PostDetailResponse": {
        "type": "object",
        "properties": {
          "postId": {
            "type": "integer",
            "format": "int64"
          },
          "content": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "mediaUrls": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "memberId": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "likeCount": {
            "type": "integer",
            "format": "int32"
          },
          "likeUsers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LikeUserDto"
            }
          },
          "likedByMe": {
            "type": "boolean"
          },
          "comments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommentListResponse"
            }
          }
        }
      },
      "PageUserPostResponse": {
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserPostResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
      "PageableObject": {
        "type": "object",
        "properties": {
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "paged": {
            "type": "boolean"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "unpaged": {
            "type": "boolean"
          }
        }
      },
      "SortObject": {
        "type": "object",
        "properties": {
          "empty": {
            "type": "boolean"
          },
          "unsorted": {
            "type": "boolean"
          },
          "sorted": {
            "type": "boolean"
          }
        }
      },
      "UserPostResponse": {
        "type": "object",
        "properties": {
          "postId": {
            "type": "integer",
            "format": "int64"
          },
          "representativeImageUrl": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "JWT": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}